feat: 임시 mock data 연결 및 알림 상태 업데이트 hook 생성
